requirements.txt
Flask==2.3.3
pandas==2.2.3
openpyxl==3.1.2
matplotlib==3.9.2
numpy==2.3.4
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
gunicorn==23.0.0
setuptools==69.5.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0